+++ src/palette/dark/background/hard.ts
@@ -4,18 +4,20 @@
  *  License:    MIT
  *--------------------------------------------------------------*/
 
+import type { Palette } from '../../../interface';
+
 export default {
-  bg0: "#0f1210",
-  bg1: "#141814",
-  bg: "#1a1f1c",
-  bg2: "#222822",
-  bg3: "#2d3830",
-  bg4: "#374139",
-  bg5: "#414b42",
-  grey0: "#7f897d",
-  grey1: "#859289",
-  grey2: "#9aa79d",
-  shadow: "#00000070",
-};
+  bg0: '#0f1210',
+  bg1: '#141814',
+  bg: '#1a1f1c',
+  bg2: '#222822',
+  bg3: '#2d3830',
+  bg4: '#374139',
+  bg5: '#414b42',
+  grey0: '#7f897d',
+  grey1: '#859289',
+  grey2: '#9aa79d',
+  shadow: '#00000070',
+} satisfies Partial<Palette>;
 
 // vim: fdm=marker fmr={{{,}}}: