+++ src/palette/dark/background/medium.ts
@@ -4,18 +4,20 @@
  *  License:    MIT
  *--------------------------------------------------------------*/
 
+import type { Palette } from '../../../interface';
+
 export default {
-  bg0: "#141814",
-  bg1: "#1f241f",
-  bg: "#222822",
-  bg2: "#2d3830",
-  bg3: "#3d4a40",
-  bg4: "#4a5a4d",
-  bg5: "#5a6a5d",
-  grey0: "#7f897d",
-  grey1: "#859289",
-  grey2: "#9aa79d",
-  shadow: "#00000070",
-};
+  bg0: '#141814',
+  bg1: '#1f241f',
+  bg: '#222822',
+  bg2: '#2d3830',
+  bg3: '#3d4a40',
+  bg4: '#4a5a4d',
+  bg5: '#5a6a5d',
+  grey0: '#7f897d',
+  grey1: '#859289',
+  grey2: '#9aa79d',
+  shadow: '#00000070',
+} satisfies Partial<Palette>;
 
 // vim: fdm=marker fmr={{{,}}}: