+++ src/palette/dark/background/soft.ts
@@ -4,18 +4,20 @@
  *  License:    MIT
  *--------------------------------------------------------------*/
 
+import type { Palette } from '../../../interface';
+
 export default {
-  bg0: "#1e2520",
-  bg1: "#242b26",
-  bg: "#2a312c",
-  bg2: "#323a34",
-  bg3: "#3c4440",
-  bg4: "#464e4a",
-  bg5: "#4e5954",
-  grey0: "#7f897d",
-  grey1: "#859289",
-  grey2: "#9aa79d",
-  shadow: "#00000070",
-};
+  bg0: '#1e2520',
+  bg1: '#242b26',
+  bg: '#2a312c',
+  bg2: '#323a34',
+  bg3: '#3c4440',
+  bg4: '#464e4a',
+  bg5: '#4e5954',
+  grey0: '#7f897d',
+  grey1: '#859289',
+  grey2: '#9aa79d',
+  shadow: '#00000070',
+} satisfies Partial<Palette>;
 
 // vim: fdm=marker fmr={{{,}}}: