+++ src/palette/light/background/medium.ts
@@ -4,18 +4,20 @@
  *  License:    MIT
  *--------------------------------------------------------------*/
 
+import type { Palette } from '../../../interface';
+
 export default {
-  bg0: "#efebd4",
-  bg1: "#f8f4e8",
-  bg: "#fdf6e3",
-  bg2: "#f4f0d9",
-  bg3: "#efebd4",
-  bg4: "#e6e2cc",
-  bg5: "#e0dcc7",
-  grey0: "#7a8478",
-  grey1: "#6b7566",
-  grey2: "#5c6658",
-  shadow: "#3c474d20",
-};
+  bg0: '#efebd4',
+  bg1: '#f8f4e8',
+  bg: '#fdf6e3',
+  bg2: '#f4f0d9',
+  bg3: '#efebd4',
+  bg4: '#e6e2cc',
+  bg5: '#e0dcc7',
+  grey0: '#7a8478',
+  grey1: '#6b7566',
+  grey2: '#5c6658',
+  shadow: '#3c474d20',
+} satisfies Partial<Palette>;
 
 // vim: fdm=marker fmr={{{,}}}: